apps/backend/pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Project Template - Python API + React frontend"
authors = [
    {name = "Adrien Carpentier", email = "me@adriencarpentier.com"},
]
dependencies = [
    "django<4.3.0,>=4.2.0",
    "django-storages[google]>=1.13.2",
    "django-environ>=0.10.0",
    "django-ninja>=0.21.0",
    "google-cloud-secret-manager>=2.16.1",
    "gunicorn>=20.1.0",
    "orjson>=3.8.14",
    "pre-commit>=3.2.2",
    "psycopg[binary]>=3.1.8",
    "python-environ>=0.4.54",
    "sentry-sdk>=1.20.0",
    "setuptools>=67.6.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "ipython>=8.12.0",
    "ruff>=0.0.263",
]

[tool.setuptools]
py-modules = []


[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
